[SOLR-13865] Migrate replica routing code to  solrJ
