+++ solr/core/src/java/org/apache/solr/handler/component/HttpShardHandlerFactory.java
@@ -16,20 +16,13 @@
  */
 package org.apache.solr.handler.component;
 
-import static org.apache.solr.util.stats.InstrumentedHttpListenerFactory.KNOWN_METRIC_NAME_STRATEGIES;
-
-import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Random;
 import java.util.Set;
@@ -41,6 +34,8 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.SynchronousQueue;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
+
+import com.google.common.annotations.VisibleForTesting;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.client.HttpClient;
 import org.apache.solr.client.solrj.SolrClient;
@@ -51,16 +46,16 @@ import org.apache.solr.client.solrj.impl.HttpClientUtil;
 import org.apache.solr.client.solrj.impl.HttpSolrClient;
 import org.apache.solr.client.solrj.impl.LBHttp2SolrClient;
 import org.apache.solr.client.solrj.impl.LBSolrClient;
-import org.apache.solr.client.solrj.impl.PreferenceRule;
+import org.apache.solr.client.solrj.routing.AffinityReplicaListTransformerFactory;
+import org.apache.solr.client.solrj.routing.ReplicaListTransformer;
+import org.apache.solr.client.solrj.routing.ReplicaListTransformerFactory;
+import org.apache.solr.client.solrj.routing.ReplicaListTransformerManager;
 import org.apache.solr.client.solrj.request.QueryRequest;
-import org.apache.solr.cloud.NodesSysPropsCacher;
 import org.apache.solr.cloud.ZkController;
 import org.apache.solr.common.SolrException;
 import org.apache.solr.common.SolrException.ErrorCode;
 import org.apache.solr.common.cloud.ClusterState;
-import org.apache.solr.common.cloud.Replica;
 import org.apache.solr.common.cloud.ZkStateReader;
-import org.apache.solr.common.params.CommonParams;
 import org.apache.solr.common.params.ShardParams;
 import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.common.util.ExecutorUtil;
@@ -83,10 +78,12 @@ import org.apache.solr.util.stats.MetricUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static org.apache.solr.util.stats.InstrumentedHttpListenerFactory.KNOWN_METRIC_NAME_STRATEGIES;
+
 public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.apache.solr.util.plugin.PluginInfoInitialized, SolrMetricProducer {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
   private static final String DEFAULT_SCHEME = "http";
-  
+
   // We want an executor that doesn't take up any resources if
   // it's not used, so it could be created statically for
   // the distributed search component if desired.
@@ -124,7 +121,7 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
 
   protected final Random r = new Random();
 
-  private final ReplicaListTransformer shufflingReplicaListTransformer = new ShufflingReplicaListTransformer(r);
+  private ReplicaListTransformerManager replicaListTransformerManager = new ReplicaListTransformerManager();
 
   // URL scheme to be used in distributed search.
   static final String INIT_URL_SCHEME = "urlScheme";
@@ -224,6 +221,8 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
 
   private void initReplicaListTransformers(NamedList routingConfig) {
     String defaultRouting = null;
+    ReplicaListTransformerFactory stableRltFactory = null;
+    ReplicaListTransformerFactory defaultRltFactory;
     if (routingConfig != null && routingConfig.size() > 0) {
       Iterator<Entry<String,?>> iter = routingConfig.iterator();
       do {
@@ -240,21 +239,22 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
           case ShardParams.REPLICA_STABLE:
             NamedList<?> c = getNamedList(e.getValue());
             defaultRouting = checkDefaultReplicaListTransformer(c, key, defaultRouting);
-            this.stableRltFactory = new AffinityReplicaListTransformerFactory(c);
+            stableRltFactory = new AffinityReplicaListTransformerFactory(c);
             break;
           default:
             throw new IllegalArgumentException("invalid replica routing spec name: " + key);
         }
       } while (iter.hasNext());
     }
-    if (this.stableRltFactory == null) {
-      this.stableRltFactory = new AffinityReplicaListTransformerFactory();
+    if (stableRltFactory == null) {
+      stableRltFactory = new AffinityReplicaListTransformerFactory();
     }
     if (ShardParams.REPLICA_STABLE.equals(defaultRouting)) {
-      this.defaultRltFactory = this.stableRltFactory;
+      defaultRltFactory = stableRltFactory;
     } else {
-      this.defaultRltFactory = this.randomRltFactory;
+      defaultRltFactory = ReplicaListTransformerManager.RANDOM_RLTF;
     }
+    this.replicaListTransformerManager = new ReplicaListTransformerManager(defaultRltFactory, stableRltFactory);
   }
 
   @Override
@@ -269,7 +269,7 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
     String strategy = getParameter(args, "metricNameStrategy", UpdateShardHandlerConfig.DEFAULT_METRICNAMESTRATEGY, sb);
     this.metricNameStrategy = KNOWN_METRIC_NAME_STRATEGIES.get(strategy);
     if (this.metricNameStrategy == null)  {
-      throw new SolrException(SolrException.ErrorCode.SERVER_ERROR,
+      throw new SolrException(ErrorCode.SERVER_ERROR,
           "Unknown metricNameStrategy: " + strategy + " found. Must be one of: " + KNOWN_METRIC_NAME_STRATEGIES.keySet());
     }
 
@@ -290,7 +290,7 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
     this.accessPolicy = getParameter(args, INIT_FAIRNESS_POLICY, accessPolicy,sb);
     this.whitelistHostChecker = new WhitelistHostChecker(args == null? null: (String) args.get(INIT_SHARDS_WHITELIST), !getDisableShardsWhitelist());
     log.info("Host whitelist initialized: {}", this.whitelistHostChecker);
-    
+
     // magic sysprop to make tests reproducible: set by SolrTestCaseJ4.
     String v = System.getProperty("tests.shardhandler.randomSeed");
     if (v != null) {
@@ -354,7 +354,7 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
         if (loadbalancer != null) {
           loadbalancer.close();
         }
-      } finally { 
+      } finally {
         if (defaultClient != null) {
           IOUtils.closeQuietly(defaultClient);
         }
@@ -409,280 +409,23 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
     return urls;
   }
 
-  /**
-   * A distributed request is made via {@link LBSolrClient} to the first live server in the URL list.
-   * This means it is just as likely to choose current host as any of the other hosts.
-   * This function makes sure that the cores are sorted according to the given list of preferences.
-   * E.g. If all nodes prefer local cores then a bad/heavily-loaded node will receive less requests from 
-   * healthy nodes. This will help prevent a distributed deadlock or timeouts in all the healthy nodes due 
-   * to one bad node.
-   * 
-   * Optional final preferenceRule is *not* used for pairwise sorting, but instead defines how "equivalent"
-   * replicas will be ordered (the base ordering). Defaults to "random"; may specify "stable".
-   */
-  static class NodePreferenceRulesComparator implements Comparator<Object> {
-
-    private final SolrQueryRequest request;
-    private final NodesSysPropsCacher sysPropsCache;
-    private final String nodeName;
-    private final List<PreferenceRule> sortRules;
-    private final List<PreferenceRule> preferenceRules;
-    private String localHostAddress = null;
-    private final ReplicaListTransformer baseReplicaListTransformer;
-
-    public NodePreferenceRulesComparator(final List<PreferenceRule> preferenceRules, final SolrQueryRequest request,
-        final ReplicaListTransformerFactory defaultRltFactory, final ReplicaListTransformerFactory randomRltFactory,
-        final ReplicaListTransformerFactory stableRltFactory) {
-      this.request = request;
-      final SolrCore core; // explicit check for null core (temporary?, for tests)
-      if (request != null && (core = request.getCore()) != null && core.getCoreContainer().getZkController() != null) {
-        ZkController zkController = request.getCore().getCoreContainer().getZkController();
-        sysPropsCache = zkController.getSysPropsCacher();
-        nodeName = zkController.getNodeName();
-      } else {
-        sysPropsCache = null;
-        nodeName = null;
-      }
-      this.preferenceRules = preferenceRules;
-      final int maxIdx = preferenceRules.size() - 1;
-      final PreferenceRule lastRule = preferenceRules.get(maxIdx);
-      if (!ShardParams.SHARDS_PREFERENCE_REPLICA_BASE.equals(lastRule.name)) {
-        this.sortRules = preferenceRules;
-        this.baseReplicaListTransformer = defaultRltFactory.getInstance(null, request, randomRltFactory);
-      } else {
-        if (maxIdx == 0) {
-          this.sortRules = null;
-        } else {
-          this.sortRules = preferenceRules.subList(0, maxIdx);
-        }
-        String[] parts = lastRule.value.split(":", 2);
-        switch (parts[0]) {
-          case ShardParams.REPLICA_RANDOM:
-            this.baseReplicaListTransformer = randomRltFactory.getInstance(parts.length == 1 ? null : parts[1], request, null);
-            break;
-          case ShardParams.REPLICA_STABLE:
-            this.baseReplicaListTransformer = stableRltFactory.getInstance(parts.length == 1 ? null : parts[1], request, randomRltFactory);
-            break;
-          default:
-            throw new IllegalArgumentException("Invalid base replica order spec");
-        }
-      }
-    }
-    private static final ReplicaListTransformer NOOP_RLT = (List<?> choices) -> { /* noop */ };
-    private static final ReplicaListTransformerFactory NOOP_RLTF = (String configSpec, SolrQueryRequest request,
-        ReplicaListTransformerFactory fallback) -> NOOP_RLT;
-    /**
-     * For compatibility with tests, which expect this constructor to have no effect on the *base* order.
-     */
-    NodePreferenceRulesComparator(final List<PreferenceRule> sortRules, final SolrQueryRequest request) {
-      this(sortRules, request, NOOP_RLTF, null, null);
-    }
-
-    public ReplicaListTransformer getBaseReplicaListTransformer() {
-      return baseReplicaListTransformer;
-    }
-
-    @Override
-    public int compare(Object left, Object right) {
-      if (this.sortRules != null) {
-        for (PreferenceRule preferenceRule: this.sortRules) {
-          final boolean lhs;
-          final boolean rhs;
-          switch (preferenceRule.name) {
-            case ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE:
-              lhs = hasReplicaType(left, preferenceRule.value);
-              rhs = hasReplicaType(right, preferenceRule.value);
-              break;
-            case ShardParams.SHARDS_PREFERENCE_REPLICA_LOCATION:
-              lhs = hasCoreUrlPrefix(left, preferenceRule.value);
-              rhs = hasCoreUrlPrefix(right, preferenceRule.value);
-              break;
-            case ShardParams.SHARDS_PREFERENCE_NODE_WITH_SAME_SYSPROP:
-              if (sysPropsCache == null) {
-                throw new IllegalArgumentException("Unable to get the NodesSysPropsCacher" +
-                    " on sorting replicas by preference:"+ preferenceRule.value);
-              }
-              lhs = hasSameMetric(left, preferenceRule.value);
-              rhs = hasSameMetric(right, preferenceRule.value);
-              break;
-            case ShardParams.SHARDS_PREFERENCE_REPLICA_BASE:
-              throw new IllegalArgumentException("only one base replica order may be specified in "
-                  + ShardParams.SHARDS_PREFERENCE + ", and it must be specified last");
-            default:
-              throw new IllegalArgumentException("Invalid " + ShardParams.SHARDS_PREFERENCE + " type: " + preferenceRule.name);
-          }
-          if (lhs != rhs) {
-            return lhs ? -1 : +1;
-          }
-        }
-      }
-      return 0;
-    }
-
-    private boolean hasSameMetric(Object o, String metricTag) {
-      if (!(o instanceof Replica)) {
-        return false;
-      }
-
-      Collection<String> tags = Collections.singletonList(metricTag);
-      String otherNodeName = ((Replica) o).getNodeName();
-      Map<String, Object> currentNodeMetric = sysPropsCache.getSysProps(nodeName, tags);
-      Map<String, Object> otherNodeMetric = sysPropsCache.getSysProps(otherNodeName, tags);
-      return currentNodeMetric.equals(otherNodeMetric);
-    }
-
-    private boolean hasCoreUrlPrefix(Object o, String prefix) {
-      final String s;
-      if (o instanceof String) {
-        s = (String)o;
-      }
-      else if (o instanceof Replica) {
-        s = ((Replica)o).getCoreUrl();
-      } else {
-        return false;
-      }
-      if (prefix.equals(ShardParams.REPLICA_LOCAL)) {
-        if (null == localHostAddress) {
-          final ZkController zkController = this.request.getCore().getCoreContainer().getZkController();
-          localHostAddress = zkController != null ? zkController.getBaseUrl() : "";
-          if (localHostAddress.isEmpty()) {
-            log.warn("Couldn't determine current host address for sorting of local replicas");
-          }
-        }
-        if (!localHostAddress.isEmpty()) {
-          if (s.startsWith(localHostAddress)) {
-            return true;
-          }
-        }
-      } else {
-        if (s.startsWith(prefix)) {
-          return true;
-        }
-      }
-      return false;
-    }
-    private static boolean hasReplicaType(Object o, String preferred) {
-      if (!(o instanceof Replica)) {
-        return false;
-      }
-      final String s = ((Replica)o).getType().toString();
-      return s.equals(preferred);
-    }
-  }
-
-  private final ReplicaListTransformerFactory randomRltFactory = (String configSpec, SolrQueryRequest request,
-      ReplicaListTransformerFactory fallback) -> shufflingReplicaListTransformer;
-  private ReplicaListTransformerFactory stableRltFactory;
-  private ReplicaListTransformerFactory defaultRltFactory;
-
-  /**
-   * Private class responsible for applying pairwise sort based on inherent replica attributes,
-   * and subsequently reordering any equivalent replica sets according to behavior specified
-   * by the baseReplicaListTransformer.
-   */
-  private static final class TopLevelReplicaListTransformer implements ReplicaListTransformer {
-
-    private final NodePreferenceRulesComparator replicaComp;
-    private final ReplicaListTransformer baseReplicaListTransformer;
-
-    public TopLevelReplicaListTransformer(NodePreferenceRulesComparator replicaComp, ReplicaListTransformer baseReplicaListTransformer) {
-      this.replicaComp = replicaComp;
-      this.baseReplicaListTransformer = baseReplicaListTransformer;
-    }
-
-    @Override
-    public void transform(List<?> choices) {
-      if (choices.size() > 1) {
-        if (log.isDebugEnabled()) {
-          log.debug("Applying the following sorting preferences to replicas: {}",
-              Arrays.toString(replicaComp.preferenceRules.toArray()));
-        }
-
-        // First, sort according to comparator rules.
-        try {
-          choices.sort(replicaComp);
-        } catch (IllegalArgumentException iae) {
-          throw new SolrException(
-              SolrException.ErrorCode.BAD_REQUEST,
-              iae.getMessage()
-          );
-        }
-
-        // Next determine all boundaries between replicas ranked as "equivalent" by the comparator
-        Iterator<?> iter = choices.iterator();
-        Object prev = iter.next();
-        Object current;
-        int idx = 1;
-        int boundaryCount = 0;
-        int[] boundaries = new int[choices.size() - 1];
-        do {
-          current = iter.next();
-          if (replicaComp.compare(prev, current) != 0) {
-            boundaries[boundaryCount++] = idx;
-          }
-          prev = current;
-          idx++;
-        } while (iter.hasNext());
-
-        // Finally inspect boundaries to apply base transformation, where necessary (separate phase to avoid ConcurrentModificationException)
-        int startIdx = 0;
-        int endIdx;
-        for (int i = 0; i < boundaryCount; i++) {
-          endIdx = boundaries[i];
-          if (endIdx - startIdx > 1) {
-            baseReplicaListTransformer.transform(choices.subList(startIdx, endIdx));
-          }
-          startIdx = endIdx;
-        }
-
-        if (log.isDebugEnabled()) {
-          log.debug("Applied sorting preferences to replica list: {}",
-              Arrays.toString(choices.toArray()));
-        }
-      }
-    }
-  }
-
   protected ReplicaListTransformer getReplicaListTransformer(final SolrQueryRequest req) {
     final SolrParams params = req.getParams();
     final SolrCore core = req.getCore(); // explicit check for null core (temporary?, for tests)
     ZkController zkController = core == null ? null : core.getCoreContainer().getZkController();
-    String defaultShardPreference = "";
     if (zkController != null) {
-      defaultShardPreference = zkController.getZkStateReader().getClusterProperties()
-          .getOrDefault(ZkStateReader.DEFAULT_SHARD_PREFERENCES, "")
-          .toString();
+      return replicaListTransformerManager.getReplicaListTransformer(
+          params,
+          zkController.getZkStateReader().getClusterProperties()
+              .getOrDefault(ZkStateReader.DEFAULT_SHARD_PREFERENCES, "")
+              .toString(),
+          zkController.getNodeName(),
+          zkController.getBaseUrl(),
+          zkController.getSysPropsCacher()
+      );
+    } else {
+      return replicaListTransformerManager.getReplicaListTransformer(params);
     }
-
-
-    @SuppressWarnings("deprecation")
-    final boolean preferLocalShards = params.getBool(CommonParams.PREFER_LOCAL_SHARDS, false);
-    final String shardsPreferenceSpec = params.get(ShardParams.SHARDS_PREFERENCE, defaultShardPreference);
-
-    if (preferLocalShards || !shardsPreferenceSpec.isEmpty()) {
-      if (preferLocalShards && !shardsPreferenceSpec.isEmpty()) {
-        throw new SolrException(
-          SolrException.ErrorCode.BAD_REQUEST,
-          "preferLocalShards is deprecated and must not be used with shards.preference" 
-        );
-      }
-      List<PreferenceRule> preferenceRules = PreferenceRule.from(shardsPreferenceSpec);
-      if (preferLocalShards) {
-        preferenceRules.add(new PreferenceRule(ShardParams.SHARDS_PREFERENCE_REPLICA_LOCATION, ShardParams.REPLICA_LOCAL));
-      }
-
-      NodePreferenceRulesComparator replicaComp = new NodePreferenceRulesComparator(preferenceRules, req,
-          defaultRltFactory, randomRltFactory, stableRltFactory);
-      ReplicaListTransformer baseReplicaListTransformer = replicaComp.getBaseReplicaListTransformer();
-      if (replicaComp.sortRules == null) {
-        // only applying base transformation
-        return baseReplicaListTransformer;
-      } else {
-        return new TopLevelReplicaListTransformer(replicaComp, baseReplicaListTransformer);
-      }
-    }
-
-    return defaultRltFactory.getInstance(null, req, randomRltFactory);
   }
 
   /**
@@ -691,7 +434,7 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
   public CompletionService newCompletionService() {
     return new ExecutorCompletionService<ShardResponse>(commExecutor);
   }
-  
+
   /**
    * Rebuilds the URL replacing the URL scheme of the passed URL with the
    * configured scheme replacement.If no scheme was configured, the passed URL's
@@ -703,7 +446,7 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
     } else if(StringUtils.isNotEmpty(scheme)) {
       return scheme + "://" + URLUtil.removeScheme(url);
     }
-    
+
     return url;
   }
 
@@ -716,28 +459,28 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
         solrMetricsContext.getMetricRegistry(),
         SolrMetricManager.mkName("httpShardExecutor", expandedScope, "threadPool"));
   }
-  
+
   /**
    * Class used to validate the hosts in the "shards" parameter when doing a distributed
    * request
    */
   public static class WhitelistHostChecker {
-    
+
     /**
      * List of the whitelisted hosts. Elements in the list will be host:port (no protocol or context)
      */
     private final Set<String> whitelistHosts;
-    
+
     /**
-     * Indicates whether host checking is enabled 
+     * Indicates whether host checking is enabled
      */
     private final boolean whitelistHostCheckingEnabled;
-    
+
     public WhitelistHostChecker(String whitelistStr, boolean enabled) {
       this.whitelistHosts = implGetShardsWhitelist(whitelistStr);
       this.whitelistHostCheckingEnabled = enabled;
     }
-    
+
     final static Set<String> implGetShardsWhitelist(final String shardsWhitelist) {
       if (shardsWhitelist != null && !shardsWhitelist.isEmpty()) {
         return StrUtils.splitSmart(shardsWhitelist, ',')
@@ -753,32 +496,32 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
                   url = new URL(hostUrl);
                 }
               } catch (MalformedURLException e) {
-                throw new SolrException(SolrException.ErrorCode.SERVER_ERROR, "Invalid URL syntax in \"" + INIT_SHARDS_WHITELIST + "\": " + shardsWhitelist, e);
+                throw new SolrException(ErrorCode.SERVER_ERROR, "Invalid URL syntax in \"" + INIT_SHARDS_WHITELIST + "\": " + shardsWhitelist, e);
               }
               if (url.getHost() == null || url.getPort() < 0) {
-                throw new SolrException(SolrException.ErrorCode.SERVER_ERROR, "Invalid URL syntax in \"" + INIT_SHARDS_WHITELIST + "\": " + shardsWhitelist);
+                throw new SolrException(ErrorCode.SERVER_ERROR, "Invalid URL syntax in \"" + INIT_SHARDS_WHITELIST + "\": " + shardsWhitelist);
               }
               return url.getHost() + ":" + url.getPort();
             }).collect(Collectors.toSet());
       }
       return null;
     }
-    
-    
+
+
     /**
      * @see #checkWhitelist(ClusterState, String, List)
      */
     protected void checkWhitelist(String shardsParamValue, List<String> shardUrls) {
       checkWhitelist(null, shardsParamValue, shardUrls);
     }
-    
+
     /**
      * Checks that all the hosts for all the shards requested in shards parameter exist in the configured whitelist
      * or in the ClusterState (in case of cloud mode)
-     * 
+     *
      * @param clusterState The up to date ClusterState, can be null in case of non-cloud mode
      * @param shardsParamValue The original shards parameter
-     * @param shardUrls The list of cores generated from the shards parameter. 
+     * @param shardUrls The list of cores generated from the shards parameter.
      */
     protected void checkWhitelist(ClusterState clusterState, String shardsParamValue, List<String> shardUrls) {
       if (!whitelistHostCheckingEnabled) {
@@ -793,7 +536,7 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
       } else {
         localWhitelistHosts = Collections.emptySet();
       }
-      
+
       shardUrls.stream().map(String::trim).forEach((shardUrl) -> {
         URL url;
         try {
@@ -804,10 +547,10 @@ public class HttpShardHandlerFactory extends ShardHandlerFactory implements org.
             url = new URL(shardUrl);
           }
         } catch (MalformedURLException e) {
-          throw new SolrException(SolrException.ErrorCode.BAD_REQUEST, "Invalid URL syntax in \"shards\" parameter: " + shardsParamValue, e);
+          throw new SolrException(ErrorCode.BAD_REQUEST, "Invalid URL syntax in \"shards\" parameter: " + shardsParamValue, e);
         }
         if (url.getHost() == null || url.getPort() < 0) {
-          throw new SolrException(SolrException.ErrorCode.BAD_REQUEST, "Invalid URL syntax in \"shards\" parameter: " + shardsParamValue);
+          throw new SolrException(ErrorCode.BAD_REQUEST, "Invalid URL syntax in \"shards\" parameter: " + shardsParamValue);
         }
         if (!localWhitelistHosts.contains(url.getHost() + ":" + url.getPort())) {
           log.warn("The '"+ShardParams.SHARDS+"' parameter value '"+shardsParamValue+"' contained value(s) not on the shards whitelist ("+localWhitelistHosts+"), shardUrl:" + shardUrl);
