+++ solr/core/src/test/org/apache/solr/handler/component/TestHttpShardHandlerFactory.java
@@ -28,7 +28,7 @@ import java.util.Set;
 
 import org.apache.solr.SolrTestCaseJ4;
 import org.apache.solr.client.solrj.impl.LBSolrClient;
-import org.apache.solr.client.solrj.impl.PreferenceRule;
+import org.apache.solr.client.solrj.routing.ReplicaListTransformer;
 import org.apache.solr.client.solrj.request.QueryRequest;
 import org.apache.solr.common.SolrException;
 import org.apache.solr.common.cloud.ClusterState;
@@ -118,208 +118,6 @@ public class TestHttpShardHandlerFactory extends SolrTestCaseJ4 {
     } finally {
       if (factory != null) factory.close();
       if (cc != null) cc.shutdown();
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  public void testNodePreferenceRulesBase() throws Exception {
-    SolrCore testCore = null;
-    HttpShardHandlerFactory fac = new HttpShardHandlerFactory();
-    fac.init(new PluginInfo(null, Collections.EMPTY_MAP));
-    SolrQueryRequestBase req;
-    NamedList<String> params = new NamedList<>();
-    List<Replica> replicas = getBasicReplicaList();
-
-    String rulesParam = ShardParams.SHARDS_PREFERENCE_REPLICA_BASE + ":stable:dividend:routingPreference";
-
-    params.add("routingPreference", "0");
-    params.add(ShardParams.SHARDS_PREFERENCE, rulesParam);
-
-    req = new SolrQueryRequestBase(testCore, params.toSolrParams()) {};
-    ReplicaListTransformer rlt = fac.getReplicaListTransformer(req);
-    rlt.transform(replicas);
-    assertEquals("node1", replicas.get(0).getNodeName());
-    assertEquals("node2", replicas.get(1).getNodeName());
-    assertEquals("node3", replicas.get(2).getNodeName());
-    req.close();
-
-    params.setVal(0, "1");
-    req = new SolrQueryRequestBase(testCore, params.toSolrParams()) {};
-    rlt = fac.getReplicaListTransformer(req);
-    rlt.transform(replicas);
-    assertEquals("node2", replicas.get(0).getNodeName());
-    assertEquals("node3", replicas.get(1).getNodeName());
-    assertEquals("node1", replicas.get(2).getNodeName());
-    req.close();
-
-    params.setVal(0, "2");
-    req = new SolrQueryRequestBase(testCore, params.toSolrParams()) {};
-    rlt = fac.getReplicaListTransformer(req);
-    rlt.transform(replicas);
-    assertEquals("node3", replicas.get(0).getNodeName());
-    assertEquals("node1", replicas.get(1).getNodeName());
-    assertEquals("node2", replicas.get(2).getNodeName());
-    req.close();
-
-    params.setVal(0, "3");
-    req = new SolrQueryRequestBase(testCore, params.toSolrParams()) {};
-    rlt = fac.getReplicaListTransformer(req);
-    rlt.transform(replicas);
-    assertEquals("node1", replicas.get(0).getNodeName());
-    assertEquals("node2", replicas.get(1).getNodeName());
-    assertEquals("node3", replicas.get(2).getNodeName());
-    req.close();
-
-    // Add a replica so that sorting by replicaType:TLOG can cause a tie
-    replicas.add(
-      new Replica(
-        "node4",
-        map(
-          ZkStateReader.BASE_URL_PROP, "http://host2_2:8983/solr",
-          ZkStateReader.NODE_NAME_PROP, "node4",
-          ZkStateReader.CORE_NAME_PROP, "collection1",
-          ZkStateReader.REPLICA_TYPE, "TLOG"
-        )
-      )
-    );
-
-    // replicaType and replicaBase combined rule param
-    rulesParam = ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE + ":NRT," + 
-      ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE + ":TLOG," + 
-      ShardParams.SHARDS_PREFERENCE_REPLICA_BASE + ":stable:dividend:routingPreference";
-
-    params.setVal(0, "0");
-    params.setVal(1, rulesParam);
-    req = new SolrQueryRequestBase(testCore, params.toSolrParams()) {};
-    rlt = fac.getReplicaListTransformer(req);
-    rlt.transform(replicas);
-    assertEquals("node1", replicas.get(0).getNodeName());
-    assertEquals("node2", replicas.get(1).getNodeName());
-    assertEquals("node4", replicas.get(2).getNodeName());
-    assertEquals("node3", replicas.get(3).getNodeName());
-    req.close();
-
-    params.setVal(0, "1");
-    req = new SolrQueryRequestBase(testCore, params.toSolrParams()) {};
-    rlt = fac.getReplicaListTransformer(req);
-    rlt.transform(replicas);
-    assertEquals("node1", replicas.get(0).getNodeName());
-    assertEquals("node4", replicas.get(1).getNodeName());
-    assertEquals("node2", replicas.get(2).getNodeName());
-    assertEquals("node3", replicas.get(3).getNodeName());
-    req.close();
-    fac.close();
-  }
-
-  @SuppressWarnings("unchecked")
-  private static List<Replica> getBasicReplicaList() {
-    List<Replica> replicas = new ArrayList<Replica>();
-    replicas.add(
-      new Replica(
-        "node1",
-        map(
-          ZkStateReader.BASE_URL_PROP, "http://host1:8983/solr",
-          ZkStateReader.NODE_NAME_PROP, "node1",
-          ZkStateReader.CORE_NAME_PROP, "collection1",
-          ZkStateReader.REPLICA_TYPE, "NRT"
-        )
-      )
-    );
-    replicas.add(
-      new Replica(
-        "node2",
-        map(
-          ZkStateReader.BASE_URL_PROP, "http://host2:8983/solr",
-          ZkStateReader.NODE_NAME_PROP, "node2",
-          ZkStateReader.CORE_NAME_PROP, "collection1",
-          ZkStateReader.REPLICA_TYPE, "TLOG"
-        )
-      )
-    );
-    replicas.add(
-      new Replica(
-        "node3",
-        map(
-          ZkStateReader.BASE_URL_PROP, "http://host2_2:8983/solr",
-          ZkStateReader.NODE_NAME_PROP, "node3",
-          ZkStateReader.CORE_NAME_PROP, "collection1",
-          ZkStateReader.REPLICA_TYPE, "PULL"
-        )
-      )
-    );
-    return replicas;
-  }
-
-  @SuppressWarnings("unchecked")
-  public void testNodePreferenceRulesComparator() throws Exception {
-    List<Replica> replicas = getBasicReplicaList();
-
-    // Simple replica type rule
-    List<PreferenceRule> rules = PreferenceRule.from(ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE + ":NRT," +
-        ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE + ":TLOG");
-    HttpShardHandlerFactory.NodePreferenceRulesComparator comparator = 
-      new HttpShardHandlerFactory.NodePreferenceRulesComparator(rules, null);
-    replicas.sort(comparator);
-    assertEquals("node1", replicas.get(0).getNodeName());
-    assertEquals("node2", replicas.get(1).getNodeName());
-
-    // Another simple replica type rule
-    rules = PreferenceRule.from(ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE + ":TLOG," +
-        ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE + ":NRT");
-    comparator = new HttpShardHandlerFactory.NodePreferenceRulesComparator(rules, null);
-    replicas.sort(comparator);
-    assertEquals("node2", replicas.get(0).getNodeName());
-    assertEquals("node1", replicas.get(1).getNodeName());
-
-    // replicaLocation rule
-    rules = PreferenceRule.from(ShardParams.SHARDS_PREFERENCE_REPLICA_LOCATION + ":http://host2:8983");
-    comparator = new HttpShardHandlerFactory.NodePreferenceRulesComparator(rules, null);
-    replicas.sort(comparator);
-    assertEquals("node2", replicas.get(0).getNodeName());
-    assertEquals("node1", replicas.get(1).getNodeName());
-
-    // Add a replica so that sorting by replicaType:TLOG can cause a tie
-    replicas.add(
-      new Replica(
-        "node4",
-        map(
-          ZkStateReader.BASE_URL_PROP, "http://host2_2:8983/solr",
-          ZkStateReader.NODE_NAME_PROP, "node4",
-          ZkStateReader.CORE_NAME_PROP, "collection1",
-          ZkStateReader.REPLICA_TYPE, "TLOG"
-        )
-      )
-    );
-
-    // replicaType and replicaLocation combined rule
-    rules = PreferenceRule.from(
-      ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE + ":NRT," + 
-      ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE + ":TLOG," + 
-      ShardParams.SHARDS_PREFERENCE_REPLICA_LOCATION + ":http://host2_2");
-    comparator = new HttpShardHandlerFactory.NodePreferenceRulesComparator(rules, null);
-    replicas.sort(comparator);
-    assertEquals("node1", replicas.get(0).getNodeName());
-    assertEquals("node4", replicas.get(1).getNodeName());
-    assertEquals("node2", replicas.get(2).getNodeName());
-    assertEquals("node3", replicas.get(3).getNodeName());
-
-    // Bad rule
-
-    try {
-      rules = PreferenceRule.from(ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE);
-      fail();
-    } catch (IllegalArgumentException e) {
-      assertEquals("Invalid shards.preference rule: " + ShardParams.SHARDS_PREFERENCE_REPLICA_TYPE, e.getMessage());
-    }
-
-    // Unknown rule
-    rules = PreferenceRule.from("badRule:test");
-    try {
-      comparator = new HttpShardHandlerFactory.NodePreferenceRulesComparator(rules, null);
-      replicas.sort(comparator);
-      fail();
-    } catch (IllegalArgumentException e) {
-      assertEquals("Invalid shards.preference type: badRule", e.getMessage());
     }
   }
